requirement.txt
langchain-ollama
langchain-community
langgraph
streamlit